GeneralQRImportMacView split
