+++ VultisigApp/VultisigApp/iOS/Components/VaultDetailScanButton+iOS.swift
@@ -10,8 +10,26 @@ import SwiftUI
 
 extension VaultDetailScanButton {
     var content: some View {
+        ZStack {
+            if ProcessInfo.processInfo.isiOSAppOnMac {
+                macContent
+            } else {
+                phoneContent
+            }
+        }
+    }
+    
+    var phoneContent: some View {
         Button {
             showSheet.toggle()
+        } label: {
+            label
+        }
+    }
+    
+    var macContent: some View {
+        NavigationLink {
+            <#code#>
         } label: {
             label
         }
